Remove assertion in getNamecallMethod
Instead of throwing an error if a namecall method can't be found, the function will now return a blank string. This seems better in my opinion because realistically this function should never fail, but even if it can/does, throwing an error seems unnecessary and could lead to other issues in production.